Refactor bot service for multi-streamer support
