+++ lib/donationalerts-poll.js
@@ -1,6 +1,7 @@
 const axios = require('axios');
 const { getStreamerDA, upsertStreamerDA, getAllStreamers, markDonationProcessed, isDonationProcessed, findUserByDAUserId, findUserByNormalizedLogin, getAvatarByTwitchId } = require('../db');
 const { emitToStreamer } = require('./bus');
+const botService = require('../services/bot');
 const { DA_CLIENT_ID, DA_CLIENT_SECRET } = require('./config');
 
 // Refresh token if needed
@@ -199,16 +200,15 @@ async function processDonation(streamerId, donation) {
     });
     
     // Добавляем аватар в активный список ПОСЛЕ отправки событий (как в команде !start)
-    const { addActiveAvatar, removeActiveAvatar } = require('../services/bot');
     try {
-      addActiveAvatar(user.twitch_user_id);
+      botService.addActiveAvatar(streamerId, user.twitch_user_id);
       console.log(`[DA Poll] Added avatar ${user.twitch_user_id} to active list for chat monitoring`);
-      
+
       // Автоматически удаляем аватар из активного списка через 5 минут
       // (аватар может остаться в overlay, но перестанет реагировать на сообщения)
       setTimeout(() => {
         try {
-          removeActiveAvatar(user.twitch_user_id);
+          botService.removeActiveAvatar(streamerId, user.twitch_user_id);
           console.log(`[DA Poll] Auto-removed avatar ${user.twitch_user_id} from active list after timeout`);
         } catch (error) {
           console.error(`[DA Poll] Error auto-removing avatar from active list: ${error.message}`);
