+++ routes/bot.js
@@ -6,14 +6,16 @@ function registerBotRoutes(app) {
     try {
       const uid = req.session.userId;
       if (!uid) return res.status(401).send('Неизвестен пользователь (нет cookie uid)');
-      
-      // Check if bot is already running
-      const botStatus = status();
+
+      const streamerKey = String(uid);
+
+      // Check if bot is already running for this streamer
+      const botStatus = status(streamerKey);
       if (botStatus.running) {
         return res.status(400).send('Бот уже подключен! Сначала остановите текущего бота.');
       }
-      
-      const { profile } = await ensureBotFor(String(uid));
+
+      const { profile } = await ensureBotFor(streamerKey);
       res.status(200).send(`✅ Бот успешно запущен и подключён к #${profile.login}. Напиши в чате "!ping" — ответит "pong".`);
     } catch (e) {
       logLine(`[bot] start error: ${e?.message || e}`);
@@ -21,9 +23,12 @@ function registerBotRoutes(app) {
     }
   });
 
-  app.post('/bot/stop', async (_req, res) => {
+  app.post('/bot/stop', async (req, res) => {
     try {
-      const changed = await stopBot();
+      const uid = req.session.userId;
+      if (!uid) return res.status(401).send('Неизвестен пользователь (нет cookie uid)');
+
+      const changed = await stopBot(String(uid));
       if (!changed) return res.status(200).send('Бот уже остановлен.');
       res.status(200).send('Бот остановлен.');
     } catch (e) {
@@ -32,8 +37,11 @@ function registerBotRoutes(app) {
     }
   });
 
-  app.get('/bot/status', (_req, res) => {
-    res.json(status());
+  app.get('/bot/status', (req, res) => {
+    const uid = req.session.userId;
+    if (!uid) return res.status(401).json({ error: 'Неизвестен пользователь' });
+
+    res.json(status(String(uid)));
   });
 }
